requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.8.2

# HTTP / Parsing
requests==2.32.3
lxml==5.3.0
beautifulsoup4==4.12.3

# Browser automation (Kabum via XHR intercept)
playwright==1.48.0
playwright-stealth==2.0.0

# Postgres
psycopg[binary]==3.2.1

# SQL Server
pyodbc==5.1.0

# Utils
tenacity==9.0.0
loguru==0.7.2
